+++ nilearn/plotting/surface/tests/conftest.py
@@ -1,10 +1,8 @@
 """Test fixtures used by nilearn.plotting.surface tests."""
 
-import numpy as np
 import pytest
 
 from nilearn._utils.helpers import is_matplotlib_installed, is_plotly_installed
-from nilearn.surface import SurfaceImage
 
 
 def pytest_generate_tests(metafunc):
@@ -36,35 +34,3 @@ def plt(request, engine):
         return request.getfixturevalue("matplotlib_pyplot")
     elif engine == "plotly":
         return request.getfixturevalue("plotly")
-
-
-@pytest.fixture
-def bg_map(rng, in_memory_mesh) -> np.ndarray:
-    """Return a background map with positive value."""
-    return np.abs(rng.standard_normal(size=in_memory_mesh.n_vertices))
-
-
-@pytest.fixture
-def surface_image_roi(surf_mask_1d) -> SurfaceImage:
-    """SurfaceImage for plotting."""
-    return surf_mask_1d
-
-
-@pytest.fixture
-def parcellation(in_memory_mesh) -> np.ndarray:
-    """Return array with parcellation with 2 regions."""
-    parcellation = np.zeros((in_memory_mesh.n_vertices,))
-    parcellation[in_memory_mesh.faces[3]] = 1
-    parcellation[in_memory_mesh.faces[5]] = 2
-    return parcellation
-
-
-@pytest.fixture
-def surface_image_parcellation(rng, in_memory_mesh) -> SurfaceImage:
-    """Return surface image with random parcellation."""
-    data = rng.integers(100, size=(in_memory_mesh.n_vertices, 1)).astype(float)
-    parcellation = SurfaceImage(
-        mesh={"left": in_memory_mesh, "right": in_memory_mesh},
-        data={"left": data, "right": data},
-    )
-    return parcellation
